refactor: Restructure code into modules for better organization and maintainability
